Don't check existing lights state when updating
I tried to reduce delay when switching multiple bulbs
by sending as few commands as possible can but it cause issues
because on the milight bridge the settings are independent -
changing the brightness when in color mode doesn't change
the brightness of white mode.

For now, do the simple stupid thing and just update all
the things every time.